requirements.txt
python-dotenv
langchain
langchain-openai
openai
PyPDF2
python-docx
reportlab